requirements.txt
streamlit
langchain-openai
langchain_core
python-dotenv
langchain_community
chromadb
